+++ MegaManLofi/IArena.h
@@ -14,7 +14,7 @@ namespace MegaManLofi
 
       virtual const std::shared_ptr<IPlayer> GetMutablePlayer() const = 0;
       virtual void SetPlayer( const std::shared_ptr<IPlayer> player ) = 0;
-      virtual void SetActiveRegion( Quad<long long> region ) = 0;
+      virtual void SetActiveRegion( Quad<float> region ) = 0;
 
       virtual const std::shared_ptr<IEntity> GetMutableEntity( int index ) const = 0;
       virtual void AddEntity( const std::shared_ptr<IEntity> entity ) = 0;
